added authorization

+++ app/Http/Controllers/IdeaController.php
@@ -16,12 +16,17 @@ class IdeaController extends Controller
     {
         $validated = request()->validate(['content' => 'required|min:5|max:240']);
 
+        $validated['user_id'] = auth()->id();
+
         $idea =  Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea created successfully.');
     }
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404, 'Youre not owner');
+        }
         $idea->delete();
 
 
@@ -29,12 +34,18 @@ class IdeaController extends Controller
     }
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404, 'Youre not owner');
+        }
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
     }
 
     public function update(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404, 'Youre not owner');
+        }
         $validated = request()->validate(['content' => 'required|min:5|max:240']);
 
         $idea->update($validated);
